requirements.txt
easyocr
opencv-python-headless
numpy
Pillow
streamlit
fuzzywuzzy
python-Levenshtein
pdf2image
reportlab
torch
torchvision
PyMuPDF
